myvenv/Scripts/requirements.txt
google-generativeai
python-dotenv
datasets
streamlit